chore: improve the complexity of some codes
